requirements.txt
Flask
groq
yt-dlp
requests
python-dotenv
gtts
pygame
streamlit